Running a queue worker will allow Scout to queue all operations that sync your model information to your search indexes, providing much better response times for your application's web interface.
I have added the QUEUE_SCOUT in my local .env environment variable and i have turned it to true, in the scout.php it is false so i haven't touched this file.
I have added a comment in the file to show that turning this to true in the .env is very important.